Reverting gc2kite PR
