+++ debian/changelog
@@ -31,10 +31,7 @@
 
   [Marco Pagani, Shreyasvi]
   * Added GMM from Bora et al. (2019)
-
-  [Marco Pagani]
   * Fixed bug in the multifault surface when defined using kite fault surfaces
-  * Added a more GC2 modular methodology for the calculation of Rx and Ry0
 
   [Giuseppina Tusa]
   * Added a new gsim file tusa_langer_azzaro_2019.py to implement the GMMs
