+++ openquake/hazardlib/geo/line.py
@@ -19,14 +19,10 @@
 """
 Module :mod:`openquake.hazardlib.geo.line` defines :class:`Line`.
 """
-import copy
-import numpy as np
+import numpy
 
 from openquake.hazardlib.geo import geodetic
 from openquake.hazardlib.geo import utils
-from openquake.hazardlib.geo import Point
-
-TOLERANCE = 0.1
 
 
 class Line(object):
@@ -47,8 +43,6 @@ class Line(object):
 
         if len(self.points) < 1:
             raise ValueError("One point needed to create a line!")
-
-        self.coo = np.array([[p.longitude, p.latitude] for p in self.points])
 
     def __eq__(self, other):
         """
@@ -76,24 +70,6 @@ class Line(object):
     def __getitem__(self, key):
         return self.points.__getitem__(key)
 
-    def flip(self):
-        """
-        Inverts the order of the points composing the line
-        """
-        self.points.reverse()
-        self.coo = np.flip(self.coo, axis=0)
-
-    @classmethod
-    def from_vectors(cls, lons, lats, deps=None):
-        """
-        Creates a line from three :numpy:`numpy.ndarray` instances containing
-        longitude, latitude and depths values
-        """
-        arrs = lons, lats
-        if deps is not None:
-            arrs = lons, lats, deps
-        return cls([Point(*coo) for coo in zip(*arrs)])
-
     def on_surface(self):
         """
         Check if this line is defined on the surface (i.e. all points
@@ -114,16 +90,6 @@ class Line(object):
         """
         return all(p.depth == self[0].depth for p in self)
 
-    def get_azimuths(self):
-        """
-        Return the azimuths of all the segments omposing the polyline
-        """
-        if len(self.points) == 2:
-            return self.points[0].azimuth(self.points[1])
-        lons = self.coo[:, 0]
-        lats = self.coo[:, 1]
-        return geodetic.azimuth(lons[:-1], lats[:-1], lons[1:], lats[1:])
-
     def average_azimuth(self):
         """
         Calculate and return weighted average azimuth of all line's segments
@@ -141,12 +107,23 @@ class Line(object):
         >>> '%.1f' % line.average_azimuth()
         '300.0'
         """
-        azimuths = self.get_azimuths()
-        lons = self.coo[:, 0]
-        lats = self.coo[:, 1]
+        if len(self.points) == 2:
+            return self.points[0].azimuth(self.points[1])
+        lons = numpy.array([point.longitude for point in self.points])
+        lats = numpy.array([point.latitude for point in self.points])
+        azimuths = geodetic.azimuth(lons[:-1], lats[:-1], lons[1:], lats[1:])
         distances = geodetic.geodetic_distance(lons[:-1], lats[:-1],
                                                lons[1:], lats[1:])
-        return get_average_azimuth(azimuths, distances)
+        azimuths = numpy.radians(azimuths)
+        # convert polar coordinates to Cartesian ones and calculate
+        # the average coordinate of each component
+        avg_x = numpy.mean(distances * numpy.sin(azimuths))
+        avg_y = numpy.mean(distances * numpy.cos(azimuths))
+        # find the mean azimuth from that mean vector
+        azimuth = numpy.degrees(numpy.arctan2(avg_x, avg_y))
+        if azimuth < 0:
+            azimuth += 360
+        return azimuth
 
     def resample(self, section_length):
         """
@@ -211,48 +188,19 @@ class Line(object):
 
         return Line(resampled_points)
 
-    def get_lengths(self) -> np.ndarray:
+    def get_length(self):
         """
-        Calculate a :class:`numpy.ndarray` instance with the length
-        of the segments composing the polyline
-
-        :returns:
-            Segments length in km.
-        """
-        lengths = []
-        for i, point in enumerate(self.points):
-            if i != 0:
-                lengths.append(point.distance(self.points[i - 1]))
-        return np.array(lengths)
-
-    def get_length(self) -> float:
-        """
-        Calculate the length of the line as a sum of lengths of all its
-        segments.
+        Calculate and return the length of the line as a sum of lengths
+        of all its segments.
 
         :returns:
             Total length in km.
         """
-        return np.sum(self.get_lengths())
-
-    def keep_corners(self, delta):
-        """
-        Removes the points where the change in direction is lower than a
-        tolerance value.
-
-        :param delta:
-            An angle in decimal degrees
-        """
-        coo = np.array([[p.longitude, p.latitude, p.depth] for p in
-                        self.points])
-        # Compute the azimuth of all the segments
-        azim = geodetic.azimuth(coo[:-1, 0], coo[:-1, 1],
-                                coo[1:, 0], coo[1:, 1])
-        pidx = set([0, coo.shape[0]-1])
-        idx = np.nonzero(np.abs(np.diff(azim)) > delta)[0]
-        pidx = sorted(list(pidx.union(set(idx+1))))
-        self.points = [Point(coo[c, 0], coo[c, 1]) for c in pidx]
-        self.coo = coo[pidx, :]
+        length = 0
+        for i, point in enumerate(self.points):
+            if i != 0:
+                length += point.distance(self.points[i - 1])
+        return length
 
     def resample_to_num_points(self, num_points):
         """
@@ -293,218 +241,3 @@ class Line(object):
             resampled_points.append(resampled)
 
         return Line(resampled_points)
-
-    def get_tu(self, mesh):
-        """
-        Computes the U and T coordinates of the GC2 method for a mesh of points
-
-        :param mesh:
-            An instance of :class:`openquake.hazardlib.geo.mesh.Mesh`
-        """
-
-        # Sites
-        mlo = mesh.lons
-        mla = mesh.lats
-
-        # Projection
-        lons = list(mlo.flatten()) + list(self.coo[:, 0])
-        lats = list(mla.flatten()) + list(self.coo[:, 1])
-        west, east, north, south = utils.get_spherical_bounding_box(lons, lats)
-        proj = utils.OrthographicProjection(west, east, north, south)
-
-        # Projected coordinates for the trace
-        tcoo = self.coo[:, 0], self.coo[:, 1]
-        txy = np.zeros_like(self.coo)
-        txy[:, 0], txy[:, 1] = proj(*tcoo)
-
-        # Projected coordinates for the sites
-        sxy = np.zeros((len(mla), 2))
-        sxy[:, 0], sxy[:, 1] = proj(mlo, mla)
-
-        # Compute u hat and t hat for each segment. tmp has shape
-        # (num_segments x 3)
-        slen, uhat, that = self.get_tu_hat()
-
-        # Lengths of the segments
-        segments_len = slen
-
-        # Get local coordinates for the sites
-        ui, ti = self.get_ui_ti(mesh, uhat, that)
-
-        # Compute the weights
-        weights, iot = get_ti_weights(ui, ti, segments_len)
-
-        # Now compute T and U
-        t_upp, u_upp = get_tu(ui, ti, segments_len, weights)
-        t_upp[iot] = 0.0
-
-        return t_upp, u_upp, weights
-
-    def get_ui_ti(self, mesh, uhat, that):
-        """
-        Compute the t and u coordinates. ti and ui have shape (num_segments x
-        num_sites)
-        """
-
-        # Creating the projection
-        if not hasattr(self, 'proj'):
-            oprj = utils.OrthographicProjection
-            proj = oprj.from_lons_lats(self.coo[:, 0], self.coo[:, 1])
-            self.proj = proj
-        proj = self.proj
-
-        # Sites projected coordinates
-        sxy = np.zeros((len(mesh.lons), 2))
-        sxy[:, 0], sxy[:, 1] = proj(mesh.lons, mesh.lats)
-
-        # Polyline projected coordinates
-        txy = np.zeros_like(self.coo)
-        txy[:, 0], txy[:, 1] = proj(self.coo[:, 0], self.coo[:, 1])
-
-        # Initializing ti and ui coordinates
-        ui = np.zeros((txy.shape[0]-1, sxy.shape[0]))
-        ti = np.zeros((txy.shape[0]-1, sxy.shape[0]))
-
-        # For each section
-        for i in range(ui.shape[0]):
-            tmp = copy.copy(sxy)
-            tmp[:, 0] -= txy[i, 0]
-            tmp[:, 1] -= txy[i, 1]
-            ui[i, :] = np.dot(tmp, uhat[i, 0:2])
-            ti[i, :] = np.dot(tmp, that[i, 0:2])
-        return ui, ti
-
-    def get_tu_hat(self):
-        """
-        Return the unit vectors defining the local origin for each segment of
-        the trace.
-
-        :param sx:
-            The vector with the x coordinates of the trace
-        :param sy:
-            The vector with the y coordinates of the trace
-        :returns:
-            Two arrays of size n x 3 (when n is the number of segments
-            composing the trace
-        """
-
-        # Creating the projection
-        if not hasattr(self, 'proj'):
-            oprj = utils.OrthographicProjection
-            proj = oprj.from_lons_lats(self.coo[:, 0], self.coo[:, 1])
-            self.proj = proj
-
-        # Projected coordinates
-        sx, sy = self.proj(self.coo[:, 0], self.coo[:, 1])
-
-        slen = ((sx[1:]-sx[:-1])**2 + (sy[1:]-sy[:-1])**2)**0.5
-        sg = np.zeros((len(sx)-1, 3))
-        sg[:, 0] = sx[1:]-sx[:-1]
-        sg[:, 1] = sy[1:]-sy[:-1]
-        uhat = get_versor(sg)
-        that = get_versor(np.cross(sg, np.array([0, 0, 1])))
-        return slen, uhat, that
-
-
-def get_average_azimuth(azimuths, distances) -> float:
-    """
-    Computes the average azimuth
-
-    :param azimuths:
-        A :class:`numpy.ndarray` instance
-    :param distances:
-        A :class:`numpy.ndarray` instance
-    :return:
-        A float with the mean azimuth in decimal degrees
-    """
-    azimuths = np.radians(azimuths)
-    # convert polar coordinates to Cartesian ones and calculate
-    # the average coordinate of each component
-    avg_x = np.mean(distances * np.sin(azimuths))
-    avg_y = np.mean(distances * np.cos(azimuths))
-    # find the mean azimuth from that mean vector
-    azimuth = np.degrees(np.arctan2(avg_x, avg_y))
-    if azimuth < 0:
-        azimuth += 360
-    return azimuth
-
-
-def get_tu(ui, ti, sl, weights):
-    """
-    Compute the T and U quantitities
-
-    :param ui:
-        A :class:`numpy.ndarray` instance of cardinality (num segments x num
-        sites)
-    :param ti:
-        A :class:`numpy.ndarray` instance of cardinality (num segments x num
-        sites)
-    :param sl:
-        A :class:`numpy.ndarray` instance with the segments' length
-    :param weights:
-        A :class:`numpy.ndarray` instance of cardinality (num segments x num
-        sites)
-    """
-    assert ui.shape == ti.shape == weights.shape
-
-    # Sum of weights - This has shape equal to the number of sites
-    weight_sum = np.sum(weights, axis=0)
-
-    # Compute T
-    t_upp = ti * weights
-    t_upp = np.divide(t_upp, weight_sum.T).T
-    t_upp = np.sum(t_upp, axis=1)
-
-    # Compute U
-    u_upp = ui[0] * weights[0]
-    for i in range(1, len(sl)):
-        delta = np.sum(sl[:i])
-        u_upp += ((ui[i] + delta) * weights[i])
-    u_upp = np.divide(u_upp, weight_sum.T).T
-    return t_upp, u_upp
-
-
-def get_ti_weights(ui, ti, segments_len):
-    """
-    Compute the weights
-    """
-    weights = np.zeros_like(ui)
-    terma = np.zeros_like(ui)
-    term1 = np.zeros_like(ui)
-    term2 = np.zeros_like(ui)
-    idx_on_trace = np.zeros_like(ui[0, :], dtype=bool)
-
-    for i in range(ti.shape[0]):
-
-        # More general case
-        cond0 = np.abs(ti[i, :]) >= TOLERANCE
-        if len(cond0):
-            terma[i, cond0] = segments_len[i] - ui[i, cond0]
-            term1[i, cond0] = np.arctan(terma[i, cond0] / ti[i, cond0])
-            term2[i, cond0] = np.arctan(-ui[i, cond0] / ti[i, cond0])
-            weights[i, cond0] = ((term1[i, cond0] - term2[i, cond0]) /
-                                 ti[i, cond0])
-
-        # Case for sites on the extension of one segment
-        cond1 = np.abs(ti[i, :]) < TOLERANCE
-        cond2 = np.logical_or(ui[i, :] < (0. - TOLERANCE),
-                              ui[i, :] > (segments_len[i] + TOLERANCE))
-        iii = np.logical_and(cond1, cond2)
-        if len(iii):
-            weights[i, iii] = 1./(ui[i, iii] - segments_len[i]) - 1./ui[i, iii]
-
-        # Case for sites on one segment
-        cond3 = np.logical_and(ui[i, :] >= (0. - TOLERANCE),
-                               ui[i, :] <= (segments_len[i] + TOLERANCE))
-        jjj = np.logical_and(cond1, cond3)
-        weights[i, jjj] = 1/(-0.01-segments_len[i])+1/0.01
-        idx_on_trace[jjj] = 1.0
-
-    return weights, idx_on_trace
-
-
-def get_versor(arr):
-    """
-    Returns the versor (i.e. a unit vector) of a vector
-    """
-    return np.divide(arr.T, np.linalg.norm(arr, axis=1)).T
