+++ openquake/hazardlib/tests/geo/line_test.py
@@ -14,15 +14,11 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 import unittest
-import numpy as np
-import matplotlib.pyplot as plt
-from openquake.hazardlib import geo
 
-PLOTTING = True
+from openquake.hazardlib import geo
 
 
 class LineResampleTestCase(unittest.TestCase):
-
     def test_resample(self):
         p1 = geo.Point(0.0, 0.0, 0.0)
         p2 = geo.Point(0.0, 0.127183341091, 14.1421356237)
@@ -155,408 +151,6 @@ class LineLengthTestCase(unittest.TestCase):
     def test(self):
         line = geo.Line([geo.Point(0, 0), geo.Point(0, 1), geo.Point(1, 2)])
         length = line.get_length()
-        expected_length = (line.points[0].distance(line.points[1]) +
-                           line.points[1].distance(line.points[2]))
+        expected_length = line.points[0].distance(line.points[1]) \
+                          + line.points[1].distance(line.points[2])
         self.assertEqual(length, expected_length)
-
-
-class LineFromVectorTest(unittest.TestCase):
-    def test(self):
-        lons = np.array([10.0, 20.0, 30.])
-        lats = np.array([40.0, 50.0, 60.])
-        line = geo.Line.from_vectors(lons, lats)
-        coo = np.array([[p.longitude, p.latitude] for p in line])
-        np.testing.assert_almost_equal(lons, coo[:, 0])
-        np.testing.assert_almost_equal(lats, coo[:, 1])
-
-    def test_with_depths(self):
-        lons = np.array([10.0, 20.0, 30.])
-        lats = np.array([40.0, 50.0, 60.])
-        deps = np.array([1.0, 2.0, 3.])
-        line = geo.Line.from_vectors(lons, lats, deps)
-        coo = np.array([[p.longitude, p.latitude, p.depth] for p in line])
-        np.testing.assert_almost_equal(lons, coo[:, 0])
-        np.testing.assert_almost_equal(lats, coo[:, 1])
-        np.testing.assert_almost_equal(deps, coo[:, 2])
-
-
-class RevertPointsTest(unittest.TestCase):
-    def test(self):
-        lons = np.array([1.0, 2.0, 3.])
-        lats = np.array([3.0, 4.0, 5.])
-        deps = np.array([6.0, 7.0, 8.])
-        line = geo.Line.from_vectors(lons, lats, deps)
-        line.flip()
-        computed = [p.longitude for p in line.points]
-        expected = [3., 2., 1.]
-        self.assertEqual(computed, expected)
-        np.testing.assert_equal(line.coo[:, 0], np.array(expected))
-
-
-class LineKeepCornersTest(unittest.TestCase):
-    def test(self):
-        #
-        #                  6 + d
-        #                   /
-        #                5 + c
-        #                 /
-        #              4 + b
-        #               /
-        #            3 + a
-        #             /
-        #  +----+----+
-        #  0    1    2
-        loa, laa = geo.geodetic.point_at(0.2, 0.0, 45., 10)  # 3
-        lob, lab = geo.geodetic.point_at(loa, laa, 55., 10)  # 4
-        loc, lac = geo.geodetic.point_at(lob, lab, 53., 10)  # 5
-        lod, lad = geo.geodetic.point_at(loc, lac, 53., 10)  # 6
-        lons = np.array([0.0, 0.1, 0.2, loa, lob, loc, lod])
-        lats = np.array([0.0, 0.0, 0.0, laa, lab, lac, lad])
-        line = geo.Line.from_vectors(lons, lats)
-        line.keep_corners(4.0)
-        coo = np.array([[p.longitude, p.latitude, p.depth] for p in line])
-        idxs = [0, 2, 3, 6]
-        expected = lons[idxs]
-        np.testing.assert_almost_equal(expected, coo[:, 0])
-        expected = lats[idxs]
-        np.testing.assert_almost_equal(expected, coo[:, 1])
-
-
-class ComputeTUTest(unittest.TestCase):
-
-    def test_compute_tu_01(self):
-        # The simplest test. Straight trace going from west to east.
-        lons = np.array([0.0, 0.1, 0.2])
-        lats = np.array([0.0, 0.0, 0.0])
-        line = geo.Line.from_vectors(lons, lats)
-        line.keep_corners(4.0)
-
-        # Prepare the mesh
-        coo = np.array([[0.2, 0.1]])
-        mesh = geo.Mesh(coo[:, 0], coo[:, 1])
-
-        # Compute the TU coordinates
-        tupp, uupp, wei = line.get_tu(mesh)
-        expected_t = geo.geodetic.distance(0.2, 0.0, 0.0,
-                                           coo[0, 0], coo[0, 1], 0.0)
-        expected_u = geo.geodetic.distance(0.0, 0.0, 0.0,
-                                           coo[0, 0], 0.0, 0.0)
-
-        # Test
-        self.assertAlmostEqual(-expected_t, tupp[0], places=4)
-        self.assertAlmostEqual(expected_u, uupp[0], places=4)
-
-    def test_compute_tu_02(self):
-
-        # Prepare the section trace
-        loa, laa = geo.geodetic.point_at(0.2, 0.0, 45., 20)
-        lons = np.array([0.0, 0.1, 0.2, loa])
-        lats = np.array([0.0, 0.0, 0.0, laa])
-        line = geo.Line.from_vectors(lons, lats)
-        line.keep_corners(4.0)
-
-        # Prepare the mesh
-        coo = np.array([[0.3, 0.0], [0.2, -0.1], [0.4, 0.3]])
-        mesh = geo.Mesh(coo[:, 0], coo[:, 1])
-
-        # Compute the TU coordinates
-        tupp, u_upp, wei = line.get_tu(mesh)
-
-        # TODO add test
-
-    def test_compute_tu_03(self):
-
-        # Prepare the section trace
-        loa, laa = geo.geodetic.point_at(0.2, 0.0, 45., 20)
-        lons = np.array([0.0, 0.1, 0.2, loa])
-        lats = np.array([0.0, 0.0, 0.0, laa])
-        line = geo.Line.from_vectors(lons, lats)
-        line.keep_corners(4.0)
-
-        # Get the site collection
-        mesh, plons, plats = get_mesh(-0.5, 1.0, -0.5, 1.0, 0.005)
-
-        # Compute the TU coordinates
-        tupp, uupp, wei = line.get_tu(mesh)
-
-        # Plotting results
-        if PLOTTING:
-            num = 10
-            z = np.reshape(tupp, plons.shape)
-            label = 'test_compute_tu_03 - T'
-            plot_pattern(lons, lats, z, plons, plats, label, num)
-            z = np.reshape(uupp, plons.shape)
-            label = 'test_compute_tu_03 - U'
-            plot_pattern(lons, lats, z, plons, plats, label, num)
-
-    def test_compute_tu_figure04(self):
-
-        # Get line
-        lons, lats, line = get_figure04_line()
-
-        # Get the mesh
-        mesh, plons, plats = get_mesh(-0.6, 0.6, -1.0, 0.4, 0.01)
-
-        # Compute the TU coordinates
-        tupp, uupp, wei = line.get_tu(mesh)
-
-        if PLOTTING:
-            num = 10
-            # T
-            z = np.reshape(tupp, plons.shape)
-            label = 'test_compute_tu_03 - T'
-            plot_pattern(lons, lats, z, plons, plats, label, num)
-            # U
-            z = np.reshape(uupp, plons.shape)
-            label = 'test_compute_tu_03 - U'
-            plot_pattern(lons, lats, z, plons, plats, label, num)
-
-
-class ComputeUiTiTest(unittest.TestCase):
-    """
-    For test 1 we use the same mapping of sites used for test 01 in
-    ComputeWeightsTest.
-    """
-
-    def test_uiti_01(self):
-
-        # Prepare the trace
-        lons = np.array([0.0, 0.2])
-        lats = np.array([0.0, 0.0])
-        line = geo.Line([geo.Point(lo, la) for lo, la in zip(lons, lats)])
-
-        # Prepare the mesh
-        coo = np.array([[0.0, 0.05], [0.2, 0.0], [0.3, -0.05],
-                        [0.1, -0.1], [0.3, 0.05]])
-        mesh = geo.Mesh(coo[:, 0], coo[:, 1])
-
-        # slen, uhat and that as expected
-        slen, uhat, that = line.get_tu_hat()
-        np.testing.assert_almost_equal(np.array([[1, 0, 0]]), uhat, decimal=5)
-
-        # Now computing ui and ti
-        ui, ti = line.get_ui_ti(mesh, uhat, that)
-
-        self.assertEqual((1, 5), ui.shape)
-        self.assertEqual((1, 5), ti.shape)
-
-        # TODO add test
-
-    def test_uiti_02(self):
-
-        # Prepare the trace
-        lons = np.array([0.0, 0.2, 0.25])
-        lats = np.array([0.0, 0.0, 0.05])
-        line = geo.Line([geo.Point(lo, la) for lo, la in zip(lons, lats)])
-
-        # Get the mesh
-        mesh, plons, plats = get_mesh(-0.4, 0.6, -0.2, 0.3, 0.005)
-
-        # slen, uhat and that as expected
-        slen, uhat, that = line.get_tu_hat()
-
-        # Now computing ui and ti
-        ui, ti = line.get_ui_ti(mesh, uhat, that)
-
-        # TODO add test
-
-        # Plot
-        if PLOTTING:
-            label = 'test_uiti_02 - ui first segment'
-            plot_pattern(lons, lats, ui[0], plons, plats, label)
-            label = 'test_uiti_02 - ui second segment'
-            plot_pattern(lons, lats, ui[1], plons, plats, label)
-            label = 'test_uiti_02 - ti first segment'
-            plot_pattern(lons, lats, ti[0], plons, plats, label)
-            label = 'test_uiti_02 - ti second segment'
-            plot_pattern(lons, lats, ti[1], plons, plats, label)
-
-
-class ComputeWeightsTest(unittest.TestCase):
-
-    def test_weights_01(self):
-        # Simple test with the following sites:
-        #
-        #       x site 0            x site 4
-        #       |
-        #       ---------------     x site 1
-        #                     |
-        #                     |
-        #                     x site 2
-        #
-        #             x site 3
-        #
-        n_seg = 1
-        n_sites = 5
-        ui = np.zeros((n_seg, n_sites))
-        ti = np.zeros((n_seg, n_sites))
-        ui[0, :] = np.array([-0.1, 30.0, 20.1, 10.0, 30.0])
-        ti[0, :] = np.array([5.0, 0.0, -5.0, -10.0, 5.0])
-        segl = np.array([20.0])
-        wei, iot = geo.line.get_ti_weights(ui, ti, segl)
-
-        # Compute weight for site 1
-        i = 0
-        sid = 1
-        expected = 1 / (ui[i, sid] - segl[i]) - 1 / ui[i, sid]
-        msg = 'Weight for site 1 is wrong'
-        self.assertAlmostEqual(expected, wei[i, sid], msg)
-
-        # Compute weight for site 2
-        i = 0
-        sid = 2
-        expected = (np.arctan((segl[i] - ui[i, sid]) / ti[i, sid]) -
-                    np.arctan(- ui[i, sid] / ti[i, sid]))
-        expected *= 1/ti[i, sid]
-        msg = 'Weight for site 2 is wrong'
-        self.assertAlmostEqual(expected, wei[i, sid], msg)
-
-    def test_weights_02(self):
-        # Check the weights for test_compute_tu_03
-
-        # Prepare the section trace
-        loa, laa = geo.geodetic.point_at(0.2, 0.0, 45., 20)
-        lons = np.array([0.0, 0.1, 0.2, loa])
-        lats = np.array([0.0, 0.0, 0.0, laa])
-        line = geo.Line.from_vectors(lons, lats)
-        line.keep_corners(4.0)
-
-        # Get the site collection
-        mesh, plons, plats = get_mesh(-0.5, 1.0, -0.5, 1.0, 0.01)
-
-        # slen, uhat and that
-        slen, uhat, that = line.get_tu_hat()
-
-        # Compute ui and ti
-        ui, ti = line.get_ui_ti(mesh, uhat, that)
-
-        # Compute weights
-        wei, iot = geo.line.get_ti_weights(ui, ti, slen)
-
-        # TODO add test
-
-        if PLOTTING:
-            label = 'test_weights_02 - weights first segment'
-            plot_pattern(lons, lats, wei[0], plons, plats, label)
-            label = 'test_weights_02 - weights second segment'
-            plot_pattern(lons, lats, wei[1], plons, plats, label)
-
-    def test_weights_figure04(self):
-
-        # Get line
-        lons, lats, line = get_figure04_line()
-
-        # Get the site collection
-        mesh, plons, plats = get_mesh(-0.2, 0.6, -0.8, 0.1, 0.0025)
-
-        # slen, uhat and that
-        slen, uhat, that = line.get_tu_hat()
-
-        # Compute ui and ti
-        ui, ti = line.get_ui_ti(mesh, uhat, that)
-
-        # Compute weights
-        wei, iot = geo.line.get_ti_weights(ui, ti, slen)
-
-        # Total weights
-        weit = np.sum(wei, axis=0)
-
-        if PLOTTING:
-            label = 'test_weights_figure04 - weights first segment'
-            plot_pattern(lons, lats, np.log10(wei[0]), plons, plats, label)
-            label = 'test_weights_figure04 - weights second segment'
-            plot_pattern(lons, lats, np.log10(wei[1]), plons, plats, label)
-            label = 'test_weights_figure04 - weights third segment'
-            plot_pattern(lons, lats, np.log10(wei[2]), plons, plats, label)
-            label = 'test_weights_figure04 - sum of weights'
-            plot_pattern(lons, lats, np.log10(weit), plons, plats, label)
-
-
-def get_figure04_line():
-    """
-    Returns the line describing the rupture trace in Figure 4 of Spudich and
-    Chiou (2015).
-
-    :returns:
-        A tuple with two :class:`numpy.ndarray` instances (longitudes and
-        latitudes of the pooints defininf the trace) and
-        :class:`openquake.hazardlib.geo.line.Line` instance
-    """
-
-    # Projection
-    oprj = geo.utils.OrthographicProjection
-    proj = oprj.from_lons_lats(np.array([-0.1, 0.1]), np.array([-0.1, 0.1]))
-
-    # Prepare the section trace
-    px = np.array([0.0, 0.0, 12.68, 35.66])
-    py = np.array([0.0, -40.0, -67.19, -86.47])
-    lons, lats = proj(px, py, reverse=True)
-    line = geo.Line.from_vectors(lons, lats)
-    return lons, lats, line
-
-
-def get_mesh(lomin, lomax, lamin, lamax, step):
-    """
-    Create a site collection
-    """
-    plons = []
-    plats = []
-    for lo in np.arange(lomin, lomax, step):
-        tlo = []
-        tla = []
-        for la in np.arange(lamin, lamax, step):
-            tlo.append(lo)
-            tla.append(la)
-        plons.append(tlo)
-        plats.append(tla)
-    plons = np.array(plons)
-    plats = np.array(plats)
-    mesh = geo.Mesh(plons.flatten(), plats.flatten())
-    return mesh, plons, plats
-
-
-def plot_pattern(lons, lats, z, plons, plats, label, num=5, show=True):
-    """
-    :param lons:
-        Traces longitudes
-    :param lats:
-        Traces latitudes
-    :param z:
-        Grid scalar quantity
-    :param plons:
-        2D Grid longitudes
-    :param plats:
-        2D Grid latitudes
-    :param scoo:
-        Sites coordinates
-    :param label:
-        Label
-    """
-
-    # If necessary, make lons and lats iterable
-    from collections.abc import Iterable
-    if len(lons) and not isinstance(lons[0], Iterable):
-        lons = [lons]
-        lats = [lats]
-
-    # Colormap
-    cmap = plt.get_cmap('coolwarm', 11)
-
-    # Plot
-    fig, ax = plt.subplots()
-    fig.set_size_inches(8, 6)
-    z = np.reshape(z, plons.shape)
-    cs = plt.contour(plons, plats, z, num, colors='k')
-    ax.clabel(cs, inline=True, fontsize=10)
-    if len(lons):
-        for los, las in zip(lons, lats):
-            ax.plot(los, las, '-r')
-            ax.plot(los[0], las[0], 'g', marker='$s$', mfc='none', ms=5)
-            ax.plot(los[-1], las[-1], 'r', marker='$e$', mfc='none', ms=5)
-    ax.scatter(plons.flatten(), plats.flatten(), c=z.flatten(), marker='.',
-               s=0.5, cmap=cmap)
-    ax.set_title(f'Test: {label}')
-    ax.axis('equal')
-    if show:
-        plt.show()
-    return ax
