--- openquake/hazardlib/tests/geo/surface/data/nzl/s1493/pro01.txt
@@ -1,2 +1,0 @@
-177.648 -37.7933 0.0 
-177.5084 -37.723 25.2
