--- openquake/hazardlib/tests/geo/surface/data/nzl/s1493/pro02.txt
@@ -1,2 +1,0 @@
-177.6565 -37.7883 0.0 
-177.5169 -37.718 25.2
