+++ openquake/hazardlib/tests/geo/surface/multi_test.py
@@ -20,40 +20,14 @@ import os
 import pathlib
 import unittest
 import numpy
-import matplotlib.pyplot as plt
 from openquake.hazardlib.geo import Point, Line
 from openquake.hazardlib.geo.mesh import Mesh
 from openquake.hazardlib.geo.geodetic import geodetic_distance
 from openquake.hazardlib.geo.surface.kite_fault import KiteSurface
 from openquake.hazardlib.geo.surface.multi import MultiSurface
-from openquake.hazardlib.tests.geo.line_test import get_mesh, plot_pattern
-from openquake.hazardlib.tests.geo.surface.kite_fault_test import plot_mesh_2d
 
 cd = pathlib.Path(__file__).parent
 aac = numpy.testing.assert_allclose
-aae = numpy.testing.assert_almost_equal
-
-PLOTTING = False
-
-
-class GetTorTestCase(unittest.TestCase):
-    """
-    Tests the extraction of the traces of the sections composing a rupture
-    """
-
-    def test_planar_tor01(self):
-        """ Planar rupture """
-        tmp = os.path.join(cd, 'data', 'msurface18.csv')
-        surf = MultiSurface.from_csv(tmp)
-        surf._set_tor()
-
-        # Expected value inferred from the input file
-        expected = numpy.array([[-117.955505, 33.769615],
-                                [-117.9903, 33.797646]])
-        aae(expected, surf.tors.lines[0].coo)
-        expected = numpy.array([[-117.9903, 33.797646],
-                                [-117.99624, 33.802425]])
-        aae(expected, surf.tors.lines[1].coo)
 
 
 class MultiSurfaceTestCase(unittest.TestCase):
@@ -83,77 +57,35 @@ class MultiSurfaceTestCase(unittest.TestCase):
         surfb = MultiSurface(surf19.surfaces + surf20.surfaces)
         rjba = surfa.get_joyner_boore_distance(mesh)[0]
         rjbb = surfb.get_joyner_boore_distance(mesh)[0]
-
-        # Test
         aac([rjba, rjbb], [85.676294, 89.225542])
 
     def test_rx(self):
-        meshtest = Mesh(numpy.array([-118.]), numpy.array([33]))  # 1 point
+        mesh = Mesh(numpy.array([-118.]), numpy.array([33]))   # 1 point
         tmp = os.path.join('data', 'msurface18.csv')
         surf18 = MultiSurface.from_csv(cd / tmp)  # 2 planes
         tmp = os.path.join('data', 'msurface19.csv')
         surf19 = MultiSurface.from_csv(cd / tmp)  # 2 planes
         tmp = os.path.join('data', 'msurface20.csv')
         surf20 = MultiSurface.from_csv(cd / tmp)  # 1 plane
-        rx18 = surf18.get_rx_distance(meshtest)[0]
-        rx19 = surf19.get_rx_distance(meshtest)[0]
-        rx20 = surf20.get_rx_distance(meshtest)[0]
+        rx18 = surf18.get_rx_distance(mesh)[0]
+        rx19 = surf19.get_rx_distance(mesh)[0]
+        rx20 = surf20.get_rx_distance(mesh)[0]
+        aac([rx18, rx19, rx20], [51.610675, 54.441119, -60.205692])
 
-        # Plotting
-        if PLOTTING:
-
-            # Creating the mesh
-            mesh, mlons, mlats = get_mesh(-118.5, -117.5, 33.0, 34., 0.005)
-
-            # Plots
-            rx = surf18.get_rx_distance(mesh)
-            _ = _plotting(surf18, rx, mlons, mlats, label='Rx - surf18')
-            plt.show()
-
-            rx = surf19.get_rx_distance(mesh)
-            _ = _plotting(surf19, rx, mlons, mlats, label='Rx - surf19')
-            plt.show()
-
-            rx = surf20.get_rx_distance(mesh)
-            _ = _plotting(surf20, rx, mlons, mlats, label='Rx - surf20')
-            plt.show()
-
-        # Test first set of surfaces
-        aac([rx18, rx19, rx20], [64.328038, 64.288793, 60.205692])
-
-        # Create surfaces and compute Rx
         surfa = MultiSurface(surf18.surfaces + surf19.surfaces)
         surfb = MultiSurface(surf19.surfaces + surf20.surfaces)
-        rxa = surfa.get_rx_distance(meshtest)[0]
-        rxb = surfb.get_rx_distance(meshtest)[0]
+        rxa = surfa.get_rx_distance(mesh)[0]
+        rxb = surfb.get_rx_distance(mesh)[0]
+        aac([rxa, rxb], [53.034889, -56.064366])
 
-        # Plotting
-        if PLOTTING:
-            mesh, mlons, mlats = get_mesh(-118.5, -117.5, 33.0, 34., 0.005)
-            rx = surfa.get_rx_distance(mesh)
-            _ = _plotting(surfa, rx, mlons, mlats, label='Rx - surfa')
-            plt.show()
-
-            rx = surfb.get_rx_distance(mesh)
-            _ = _plotting(surfb, rx, mlons, mlats, label='Rx - surfb')
-            plt.show()
-
-        # Test second set of surfaces
-        aac([rxa, rxb], [64.309214, 62.332508], rtol=1e-5)
-
-    def test_rx_ry0_kite(self):
-
-        # Define the surface that is a plane dipping towards north
+    def test_rx_kite(self):
         spc = 2.0
         pro1 = Line([Point(0.2, 0.0, 0.0), Point(0.2, 0.05, 15.0)])
         pro2 = Line([Point(0.0, 0.0, 0.0), Point(0.0, 0.05, 15.0)])
         sfc1 = KiteSurface.from_profiles([pro1, pro2], spc, spc)
         msurf = MultiSurface([sfc1])
-
-        # Define the mesh of sites
         pcoo = numpy.array([[0.2, 0.1], [0.0, -0.1]])
         mesh = Mesh(pcoo[:, 0], pcoo[:, 1])
-
         # Compute expected distances
         lo = pro1.points[0].longitude
         la = pro1.points[0].longitude
@@ -161,47 +93,7 @@ class MultiSurfaceTestCase(unittest.TestCase):
         lo = pro2.points[0].longitude
         la = pro2.points[0].longitude
         tmp1 = geodetic_distance(lo, la, pcoo[1, 0], pcoo[1, 1])
-
-        # Checking Rx
+        # Checking
         rx = msurf.get_rx_distance(mesh)
-        expected = numpy.array([-tmp0, tmp1])
-        computed = numpy.squeeze(rx)
-        numpy.testing.assert_almost_equal(expected, computed, decimal=5)
-
-        # Checking Ry0
-        expected = numpy.array([0.0, tmp1])
-        ry0 = msurf.get_ry0_distance(mesh)
-
-        # Plotting
-        if PLOTTING:
-            mesh, mlons, mlats = get_mesh(-0.2, 0.4, -0.2, 0.2, 0.0025)
-            rx = msurf.get_rx_distance(mesh)
-            ax = _plotting(msurf, rx, mlons, mlats, label='Rx - Kite')
-            ax.plot(pcoo[:, 0], pcoo[:, 1], 'o')
-            for srf in msurf.surfaces:
-                plot_mesh_2d(ax, srf)
-            plt.show()
-
-            ry0 = msurf.get_ry0_distance(mesh)
-            ax = _plotting(msurf, ry0, mlons, mlats, label='Ry0 - Kite')
-            ax.plot(pcoo[:, 0], pcoo[:, 1], 'o')
-            for srf in msurf.surfaces:
-                plot_mesh_2d(ax, srf)
-            plt.show()
-
-
-def _plotting(surf, dst, mlons, mlats, lons=[], lats=[], label=''):
-    """
-    Plots mesh and surface
-    """
-    num = 10
-    ax = plot_pattern(lons, lats, numpy.reshape(dst, mlons.shape),
-                      mlons, mlats, label, num, show=False)
-    lons = surf.surfaces[0].mesh.lons
-    lats = surf.surfaces[0].mesh.lats
-    for line in surf.tors.lines:
-        ax.plot(line.coo[:, 0], line.coo[:, 1], '-r')
-        ax.plot(line.coo[0, 0], line.coo[0, 1], 'g', marker="$s$", ms=5)
-        ax.plot(line.coo[-1, 0], line.coo[-1, 1], 'r', marker="$e$",
-                mfc='none', ms=8)
-    return ax
+        expected = numpy.array([tmp0, -tmp1])
+        numpy.testing.assert_almost_equal(expected, rx, decimal=5)
