--- openquake/hazardlib/tests/geo/surface/results/results_nzl_2_mesh_lons.txt
@@ -1,10 +1,0 @@
-1.670845586257339903e+02 1.670626282294073519e+02 1.670394996641763328e+02 1.670129954216545229e+02 1.669821000000000026e+02
-1.670816287555295503e+02 1.670596983508229414e+02 1.670365697784991426e+02 1.670100655303696442e+02 1.669791701063952303e+02
-1.670786992234471882e+02 1.670567688105542175e+02 1.670336402313211011e+02 1.670071359777246585e+02 1.669762405514884165e+02
-1.670757700294166170e+02 1.670538396085307511e+02 1.670307110225717224e+02 1.670042067636489662e+02 1.669733113352089333e+02
-1.670728411733673227e+02 1.670509107446820565e+02 1.670277821521804356e+02 1.670012778880720532e+02 1.669703824574862949e+02
-1.670699126552290181e+02 1.670479822189377330e+02 1.670248536200767546e+02 1.669983493509233767e+02 1.669674539182498734e+02
-1.670669844749314166e+02 1.670450540312273517e+02 1.670219254261902790e+02 1.669954211521324225e+02 1.669645257174291260e+02
-1.670640566324040321e+02 1.670421261814805405e+02 1.670189975704504661e+02 1.669924932916286195e+02 1.669615978549535384e+02
-1.670611291275766064e+02 1.670391986696268702e+02 1.670160700527869153e+02 1.669895657693415103e+02 1.669586703307526250e+02
-1.670582019603788524e+02 1.670362714955959973e+02 1.670131428731291692e+02 1.669866385852006658e+02 1.669557431447559281e+02
