--- openquake/hazardlib/tests/geo/surface/results/results_nzl_2_ref_idx.txt
@@ -1,1 +1,0 @@
-0.000000000000000000e+00
