--- openquake/hazardlib/tests/geo/surface/results/results_nzl_2_rprof_0.txt
@@ -1,10 +1,0 @@
-1.669821000000000026e+02 -4.503929999999999723e+01 0.000000000000000000e+00
-1.669791701063952303e+02 -4.503599228637377649e+01 2.462056629410059205e+00
-1.669762405514884165e+02 -4.503268449784434324e+01 4.924113258820125516e+00
-1.669733113352089333e+02 -4.502937663442897787e+01 7.386169888230158520e+00
-1.669703824574862949e+02 -4.502606869614499630e+01 9.848226517640201294e+00
-1.669674539182498734e+02 -4.502276068300970024e+01 1.231028314705020321e+01
-1.669645257174291260e+02 -4.501945259504035590e+01 1.477233977646016427e+01
-1.669615978549535384e+02 -4.501614443225427920e+01 1.723439640587016441e+01
-1.669586703307526250e+02 -4.501283619466874342e+01 1.969645303528019298e+01
-1.669557431447559281e+02 -4.500952788230102186e+01 2.215850966469022865e+01
