--- openquake/hazardlib/tests/geo/surface/results/results_nzl_2_rprof_2.txt
@@ -1,10 +1,0 @@
-1.670461999999999989e+02 -4.501930000000000121e+01 0.000000000000000000e+00
-1.670432701160867168e+02 -4.501599229616804365e+01 2.462049339280850457e+00
-1.670403405706329067e+02 -4.501268451743329280e+01 4.924098678561678710e+00
-1.670374113635680544e+02 -4.500937666381306457e+01 7.386148017842502078e+00
-1.670344824948216456e+02 -4.500606873532465357e+01 9.848197357123289919e+00
-1.670315539643232228e+02 -4.500276073198534021e+01 1.231024669640406266e+01
-1.670286257720023286e+02 -4.499945265381241200e+01 1.477229603568482474e+01
-1.670256979177885057e+02 -4.499614450082314221e+01 1.723434537496553176e+01
-1.670227704016113535e+02 -4.499283627303481836e+01 1.969639471424617128e+01
-1.670198432234003860e+02 -4.498952797046471375e+01 2.215844405352664381e+01
