+++ openquake/qa_tests_data/classical/case_61/expected/hcurve-mean.csv
@@ -1,3 +1,3 @@
-#,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,"generated_by='OpenQuake engine 3.13.0-gita7f2760ee6', start_date='2022-01-18T10:21:50', checksum=1758247745, kind='mean', investigation_time=1.0, imt='SA(0.2)'"
+#,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,"generated_by='OpenQuake engine 3.13.0-git8f2eedb69b', start_date='2022-01-12T18:16:54', checksum=1758247745, kind='mean', investigation_time=1.0, imt='SA(0.2)'"
 lon,lat,depth,poe-0.0050000,poe-0.0057376,poe-0.0065840,poe-0.0075552,poe-0.0086698,poe-0.0099487,poe-0.0114163,poe-0.0131004,poe-0.0150330,poe-0.0172506,poe-0.0197953,poe-0.0227155,poe-0.0260664,poe-0.0299117,poe-0.0343242,poe-0.0393876,poe-0.0451980,poe-0.0518655,poe-0.0595165,poe-0.0682963,poe-0.0783711,poe-0.0899323,poe-0.1031988,poe-0.1184225,poe-0.1358919,poe-0.1559383,poe-0.1789419,poe-0.2053390,poe-0.2356300,poe-0.2703896,poe-0.3102768,poe-0.3560480,poe-0.4085713,poe-0.4688427,poe-0.5380052,poe-0.6173704,poe-0.7084434,poe-0.8129511,poe-0.9328756,poe-1.0704911,poe-1.2284072,poe-1.4096188,poe-1.6175622,poe-1.8561809,poe-2.1300000
-1.00000,0.10000,0.00000,3.989106E-02,3.989106E-02,3.989106E-02,3.989106E-02,3.989106E-02,3.989106E-02,3.989106E-02,3.989099E-02,3.988936E-02,3.988292E-02,3.986610E-02,3.982966E-02,3.975823E-02,3.962383E-02,3.939355E-02,3.902447E-02,3.846508E-02,3.765408E-02,3.652752E-02,3.502796E-02,3.311520E-02,3.077742E-02,2.804027E-02,2.497123E-02,2.167729E-02,1.829494E-02,1.497405E-02,1.185826E-02,9.066324E-03,6.678345E-03,4.729711E-03,3.213289E-03,2.088316E-03,1.293031E-03,7.582298E-04,4.168123E-04,2.105508E-04,9.351457E-05,3.273851E-05,7.920155E-06,6.220758E-07,0.000000E+00,0.000000E+00,0.000000E+00,0.000000E+00
+1.00000,0.10000,0.00000,3.989106E-02,3.989106E-02,3.989106E-02,3.989106E-02,3.989106E-02,3.989106E-02,3.989106E-02,3.989097E-02,3.988924E-02,3.988255E-02,3.986524E-02,3.982791E-02,3.975504E-02,3.961845E-02,3.938504E-02,3.901166E-02,3.844669E-02,3.762887E-02,3.649452E-02,3.498669E-02,3.306587E-02,3.072106E-02,2.797873E-02,2.490702E-02,2.161325E-02,1.823392E-02,1.491851E-02,1.180998E-02,9.026247E-03,6.646594E-03,4.705706E-03,3.195972E-03,2.076399E-03,1.285220E-03,7.533882E-04,4.140018E-04,2.090367E-04,9.279467E-05,3.246561E-05,7.846078E-06,6.146182E-07,0.000000E+00,0.000000E+00,0.000000E+00,0.000000E+00
