+++ openquake/qa_tests_data/classical/case_62/expected/hcurve-mean.csv
@@ -1,3 +1,3 @@
-#,,,,,,,,,,,,,,,,,,,,,,"generated_by='OpenQuake engine 3.13.0-gitbcdcbca587', start_date='2022-01-18T10:38:53', checksum=2332671777, kind='mean', investigation_time=1.0, imt='PGA'"
+#,,,,,,,,,,,,,,,,,,,,,,"generated_by='OpenQuake engine 3.13.0-git47489c4c0b', start_date='2021-11-05T10:41:32', checksum=2332671777, kind='mean', investigation_time=1.0, imt='PGA'"
 lon,lat,depth,poe-0.0050000,poe-0.0070015,poe-0.0098041,poe-0.0137286,poe-0.0192240,poe-0.0269192,poe-0.0376948,poe-0.0527836,poe-0.0739125,poe-0.1034991,poe-0.1449289,poe-0.2029427,poe-0.2841789,poe-0.3979333,poe-0.5572227,poe-0.7802743,poe-1.0926116,poe-1.5299748,poe-2.1424109,poe-3.0000000
-86.00000,37.00000,0.00000,7.374808E-03,7.295491E-03,7.047554E-03,6.460668E-03,5.408780E-03,3.980050E-03,2.508761E-03,1.354283E-03,6.518297E-04,3.017285E-04,1.394041E-04,6.289418E-05,2.645243E-05,8.490015E-06,1.870199E-06,4.648973E-08,0.000000E+00,0.000000E+00,0.000000E+00,0.000000E+00
+86.00000,37.00000,0.00000,7.373493E-03,7.290650E-03,7.034147E-03,6.432725E-03,5.364975E-03,3.928416E-03,2.463017E-03,1.323839E-03,6.366128E-04,2.960178E-04,1.377953E-04,6.289418E-05,2.645243E-05,8.490015E-06,1.870199E-06,4.648973E-08,0.000000E+00,0.000000E+00,0.000000E+00,0.000000E+00
