+++ openquake/qa_tests_data/classical/case_65/expected/hcurve-mean.csv
@@ -1,3 +1,3 @@
-#,,,,,,,,,,,,,,,,,,,,,,"generated_by='OpenQuake engine 3.13.0-git1da0589f16', start_date='2022-01-06T22:25:08', checksum=1902944982, kind='mean', investigation_time=1.0, imt='PGA'"
+#,,,,,,,,,,,,,,,,,,,,,,"generated_by='OpenQuake engine 3.12.0-gitd508d27de0', start_date='2021-04-24T17:18:36', checksum=1037692201, kind='mean', investigation_time=1.0, imt='PGA'"
 lon,lat,depth,poe-0.0050000,poe-0.0070015,poe-0.0098041,poe-0.0137286,poe-0.0192240,poe-0.0269192,poe-0.0376948,poe-0.0527836,poe-0.0739125,poe-0.1034991,poe-0.1449289,poe-0.2029427,poe-0.2841789,poe-0.3979333,poe-0.5572227,poe-0.7802743,poe-1.0926116,poe-1.5299748,poe-2.1424109,poe-3.0000000
 9.85000,45.00000,0.00000,4.955378E-01,4.944550E-01,4.916340E-01,4.855190E-01,4.744818E-01,4.578598E-01,4.368249E-01,4.139062E-01,3.907225E-01,3.651617E-01,3.306468E-01,2.799250E-01,2.124810E-01,1.389975E-01,7.591310E-02,3.385009E-02,1.215109E-02,3.483441E-03,7.947725E-04,1.443862E-04
