+++ openquake/qa_tests_data/ucerf/expected/hazard_map-mean.csv
@@ -1,4 +1,4 @@
-#,,,"generated_by='OpenQuake engine 3.13.0-gitcd193d6ad5', start_date='2022-01-06T22:56:44', checksum=1692927904, kind='mean', investigation_time=50.0"
+#,,,"generated_by='OpenQuake engine 3.11.0-git6f08fb75b7', start_date='2021-02-04T05:35:55', checksum=3715445972, kind='mean', investigation_time=50.0"
 lon,lat,PGA-0.02,PGA-0.2
 -123.01707,38.81340,1.003430E-01,0.000000E+00
 -122.91556,38.51995,1.002583E-01,0.000000E+00
@@ -64,7 +64,7 @@ lon,lat,PGA-0.02,PGA-0.2
 -122.24958,37.84817,1.065442E-01,0.000000E+00
 -122.24067,37.73345,2.003020E-01,0.000000E+00
 -122.24059,38.13660,2.067432E-01,0.000000E+00
--122.23652,38.09696,1.081251E-01,0.000000E+00
+-122.23652,38.09696,1.100932E-01,0.000000E+00
 -122.22578,38.08201,2.115720E-01,0.000000E+00
 -122.22484,38.32268,1.100932E-01,0.000000E+00
 -122.22160,38.11647,1.081908E-01,0.000000E+00
